Dynamic View Controller
This project is unfinished. I was only able to have the image display 1 image when you click 1 and 2 images when you click 2 but I wasn't able to have the program display 3, 4, 5, and 6 images. I believe that it's because of the constraints. I have them commented out in this program because the program wouldn't run with it in there. I tried a bunch of different things but nothing worked :/. If you don't mind, I would love some help whenever you can.